requirements.txt
tensorboard==2.9.1
scikit-learn==1.0.2
toma==1.1.0
matplotlib==3.5.2
tqdm==4.63.0
pyro-ppl==1.8.1
torchmetrics==0.11.0
cityscapesScripts==2.2.2
pytorch==1.13.1
numpy==1.24.4
pillow==10.4.0
tensorboard==2.9.1
